requirements.txt
altgraph==0.17.3
buildozer==1.5.0
certifi==2022.12.7
charset-normalizer==3.1.0
distlib==0.3.6
docutils==0.19
filelock==3.10.7
idna==3.4
Kivy==2.1.0
Kivy-examples==2.1.0
Kivy-Garden==0.1.5
kivymd==1.1.1
macholib==1.16.2
mysql-connector-python==8.0.32
networkx==3.0
pexpect==4.8.0
Pillow==9.4.0
platformdirs==3.2.0
protobuf==3.20.3
ptyprocess==0.7.0
Pygments==2.14.0
pyinstaller==5.10.0
pyinstaller-hooks-contrib==2023.2
requests==2.28.2
sh==2.0.3
urllib3==1.26.15
virtualenv==20.21.0
watchdog==3.0.0
